+++ src/main/java/com/ntapia/bank/dao/CustomerRepository.java
@@ -1,4 +1,6 @@
-package com.ntapia.bank.customer;
+package com.ntapia.bank.dao;
+
+import com.ntapia.bank.model.Customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -7,7 +9,7 @@ import java.util.Optional;
 /**
  * Customer Repository
  */
-interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     Optional<Customer> findByFullName(String fullName);
 }
