+++ src/main/java/com/ntapia/bank/exception/CustomerAlreadyExistException.java
@@ -1,5 +1,4 @@
-// Copyright (c) 2018 Boomi, Inc.
-package com.ntapia.bank.customer;
+package com.ntapia.bank.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -8,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  *
  */
 @ResponseStatus(value = HttpStatus.ALREADY_REPORTED)
-class CustomerAlreadyExistException extends RuntimeException {
+public class CustomerAlreadyExistException extends RuntimeException {
 
     public CustomerAlreadyExistException() {
         super("Customer already exist!");
