+++ src/main/java/com/ntapia/bank/exception/CustomerException.java
@@ -1,5 +1,4 @@
-// Copyright (c) 2018 Boomi, Inc.
-package com.ntapia.bank.customer;
+package com.ntapia.bank.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -8,8 +7,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  *
  */
 @ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
-class CustomerException extends RuntimeException {
+public class CustomerException extends RuntimeException {
 
     public CustomerException(String message) {
+        super(message);
     }
 }
